[docs]: fix Versions page GitHub api issue
Resolves #14762
